Move contextual balloon related utility methods to utils/ui/contextualballoon.js.
