+++ packages/ckeditor5-table/src/tablecellproperties/tablecellpropertiesui.js
@@ -15,13 +15,11 @@ import TableCellPropertiesView from './ui/tablecellpropertiesview';
 import tableCellProperties from './../../theme/icons/table-cell-properties.svg';
 import {
 	colorFieldValidator,
-	getBalloonCellPositionData,
 	getLocalizedColorErrorText,
 	getLocalizedLengthErrorText,
 	defaultColors,
 	lengthFieldValidator,
-	lineWidthFieldValidator,
-	repositionContextualBalloon
+	lineWidthFieldValidator
 } from '../ui/utils';
 import {
 	getLocalizedColorOptions,
@@ -29,6 +27,7 @@ import {
 } from '@ckeditor/ckeditor5-ui/src/colorgrid/utils';
 import { debounce } from 'lodash-es';
 import { getTableWidgetAncestor } from '../utils/ui/widget';
+import { getBalloonCellPositionData, repositionContextualBalloon } from '../utils/ui/contextualballoon';
 
 const ERROR_TEXT_TIMEOUT = 500;
